[mod] data jdbcは複合主キーに対応していないため、jdbcTemplateを用いるように変更

+++ docker-local/mysql/initdb.d/create-tables.sql
@@ -1,13 +1,14 @@
 drop table if exists user_t;
 create table if not exists user_t(
-    `user_id` bigint not null auto_increment,
+    `user_id` int not null auto_increment,
     `user_name` varchar(255) not null,
     primary key (user_id)
 );
 
 drop table if exists user_role_t;
 create table if not exists user_role_t(
-     `user_id` bigint not null,
+     `user_id` int not null,
      `role_type` varchar(255) not null,
      primary key (user_id, role_type)
+--      unique ui_user_role_t_01 (user_id, role_type)
 );
